save before trying delete package lock json
